Move service-events tests to top-level test/app_signals_service_events dir
Match the repo convention of flat, sibling test directories under test/
(e.g. emf, emf_concurrent) instead of nesting under app_signals/. Keeps the
package separate from the old app_signals test_runner suite so CI runs only
these tests.
