+++ test/app_signals_service_events/agent_configs/config_logs_dynamic_placeholders.json
