+++ test/app_signals_service_events/agent_configs/credentials
