+++ test/app_signals_service_events/resources/metrics/server_consumer.json
