+++ test/app_signals_service_events/resources/send_otlp_metrics.sh
